分享页面动态数据绑定

+++ app/Http/Controllers/API/ActivitiesCollectController.php
@@ -7,12 +7,18 @@ use App\Http\Controllers\Controller;
 use App\Models\ActivitiesCollect;
 use App\Libraries\Common;
 use UUID;
+use App\Models\Activities;
+
 class ActivitiesCollectController extends Controller
 {
     /**
      * 收藏活动
      */
     public function store(Request $request){
+        $act = Activities::find($request->input('activities_id'));
+        if(!$act)
+            return Common::returnResult(204,'该活动信息不存在',"");
+
     	$collect = ActivitiesCollect::where('activities_id',$request->input('activities_id'))->where('user_id',$request->input('users_id'))->first();
     	if(!$collect){
     		$collect = new ActivitiesCollect();
@@ -20,9 +26,11 @@ class ActivitiesCollectController extends Controller
     		$collect->activities_id = $request->input('activities_id');
     		$collect->user_id = $request->input('users_id');
     		if($collect->save()){
-    			return Common::returnSuccessResult(200,'关注成功','');
+                $act->collect_count = (int)$act->collect_count +1;
+                $act->save();
+    			return Common::returnSuccessResult(200,'收藏成功','');
     		}else{
-    			return Common::returnErrorResult(400,'关注失败');
+    			return Common::returnErrorResult(400,'收藏失败');
     		}
     	}else{
     		return Common::returnErrorResult(201,'您已收藏该活动');
@@ -33,9 +41,16 @@ class ActivitiesCollectController extends Controller
      * 取消收藏
      */
     public function destory(Request $request){
+         $act = Activities::find($request->input('activities_id'));
+        if(!$act)
+            return Common::returnResult(204,'该活动信息不存在',"");
+
     	$collect = ActivitiesCollect::where('activities_id',$request->input('activities_id'))->where('user_id',$request->input('users_id'))->first();
     	if($collect){
     		if($collect->delete()){
+                $act->collect_count = (int)$act->collect_count -1;
+                $act->save();
+
     			return Common::returnSuccessResult(200,'取消成功','');
     		}
     		else{
